2 v4

+++ python_async_function/2-measure_runtime.py
@@ -10,7 +10,7 @@ import asyncio
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: float) -> float:
+def measure_time(n: int, max_delay: int) -> float:
     """Use the time module to measure
     an approximate elapsed time."""
     start_time: float = time.time()
